refactor: move SwitchThemeButton widget to core folder
